task-3736: refactor existing Validation Result
