+++ rosetta-runtime/src/main/java/com/rosetta/model/lib/meta/Key.java
@@ -18,7 +18,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 import java.util.function.Function;
-import com.rosetta.model.lib.validation.ValidationResult.ValidationType;
+import com.rosetta.model.lib.validation.ValidationType;
 
 /**
  * @author TomForwood
