+++ rosetta-runtime/src/main/java/com/rosetta/model/lib/validation/ValidationType.java
@@ -1,0 +1,6 @@
+package com.rosetta.model.lib.validation;
+
+// @Compat: MODEL_INSTANCE is replaced by CARDINALITY, TYPE_FORMAT, KEY and can be removed in the future.
+public enum ValidationType {
+    DATA_RULE, CHOICE_RULE, MODEL_INSTANCE, CARDINALITY, TYPE_FORMAT, KEY, ONLY_EXISTS, PRE_PROCESS_EXCEPTION, POST_PROCESS_EXCEPTION
+}
